requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0
requests>=2.31.0
flask>=2.3.0
flask-cors>=4.0.0
plotly>=5.15.0
matplotlib>=3.7.0
dataclasses-json>=0.6.0
python-multipart>=0.0.6
